% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sst_from_date.R
\name{extract_sst_from_date}
\alias{extract_sst_from_date}
\title{Title}
\usage{
extract_sst_from_date(date, dataset, long, lat, ...)
}
\arguments{
\item{date}{Date}

\item{dataset}{Dataset}

\item{long}{Longitude}

\item{lat}{Latitude}

\item{...}{Additional arguments}
}
\value{
Mean sea surface temperature
}
\description{
Title
}
