% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hadisst1.R
\name{read_hadsst1}
\alias{read_hadsst1}
\title{Title}
\usage{
read_hadsst1(filename, ...)
}
\arguments{
\item{filename}{Path of netCDF file}

\item{...}{Arguments passed to raster::brick()}
}
\value{
RasterBrick object
}
\description{
Title
}
