% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hurdat2.R
\name{read_hurdat2}
\alias{read_hurdat2}
\title{Read HURDAT2 dataset}
\usage{
read_hurdat2(con, interp_wind = FALSE)
}
\arguments{
\item{con}{a connection object or a character string. This is the file path or URL for the HURDAT2 dataset to read.}

\item{interp_wind}{whether missing wind speeds should be linearly interpolated.}
}
\value{
Hurricane observations dataframe
}
\description{
Read HURDAT2 dataset
}
