% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oisst.R
\name{read_oisst}
\alias{read_oisst}
\title{Read OISST dataset}
\usage{
read_oisst(filename, landsea_mask, long_w, long_e, lat_s, lat_n)
}
\arguments{
\item{filename}{full path to unzipped netCDF data file}

\item{landsea_mask}{full path to land-sea mask netCDF file}

\item{long_w}{western-most longitude of search area, must be smaller than long_e}

\item{long_e}{(optional) eastern-most longitude of search area, must be larger than long_w}

\item{lat_s}{southern-most latitude of search area, must be smaller than lat_n}

\item{lat_n}{(optional) northern-most latitude of search area, must be larger than lat_s}
}
\value{
A 2-dimensional matrix with latitudes in rows and longitudes in columns
}
\description{
Read and extract mean sea surface temperature data from NOAA's Optimum Interpolated Sea Surface Temperature (OISST) v2 High Resolution daily or weekly datasets.
}
